+++ src/renderer/views/EnvironmentView.tsx
@@ -3,7 +3,7 @@ import { motion } from 'framer-motion';
 import { useParams } from 'react-router';
 import { useTranslation } from 'react-i18next';
 import globalContainerVariants from '../utils/globalContainerVariants';
-import { getEnvironmentById } from '../ipc/ipcHandler';
+import { getEnvironmentById } from '../ipc/environmentsIpcHandler';
 import EnvironmentViewParams from '../../common/interfaces/EnvironmentViewParams';
 import serverImg from '../assets/img/server.png';
 import '../assets/styles/components/CenterView.scss';